Made Window a singleton to avoid issues with multiple windows

+++ GameEngineC++/GameEngineC++/listener.cpp
@@ -16,8 +16,8 @@ void Listener::key_callback(GLFWwindow* window, int key, int scancode, int actio
 
 /* Mouse listener */
 void Listener::cursor_position_callback(GLFWwindow* window, double xPos, double yPos) {
-	Listener::getInstance().mPos.x = xPos;
-	Listener::getInstance().mPos.y = yPos;
+	Listener::getInstance().mPos.x = (float)xPos;
+	Listener::getInstance().mPos.y = (float)yPos;
 }
 
 void Listener::mouse_button_callback(GLFWwindow* window, int button, int action, int mods) {
